pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zt-worker"
version = "0.1.0"
description = "Zero Trust Worker CLI - employee-agent container supervisor"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "docker>=7.0",
    "rich>=13.0",
]

[project.scripts]
zt-worker = "worker_cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["worker_cli*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
